pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsml-kit"
version = "0.2.0"
description = "Profile-based ML workspace with Docker Compose + uv, managed via a simple CLI."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Mihnea Teodor Stoica" }
]
dependencies = [
  "docker>=7.1",
  "typer>=0.12",
  "rich>=13",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.4",
]

[project.scripts]
dsml = "dsml.cli:app"

[dependency-groups]
dev = [
  "dsml-kit",
  "pytest>=7.4.4",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dsml = ["templates/*.yml"]

[tool.setuptools.data-files]
"share/dsml-kit/profiles" = [
  "profiles/base.yml",
  "profiles/dev.yml",
  "profiles/extended.yml",
  "profiles/minimal.yml",
  "profiles/gpu.yml",
  "profiles/full.yml",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
dsml-kit = { workspace = true }
